Extract SyncState class to its own file
